Reorganize Upload errors
Problem: upload error is defined inside Service.hs and increases its
size. Also some possible errors are defined in Edna.Web.Types. There
we have EdnaServerError which is supposed to group all errors, but
we are going to use structure of modules where each domain has
Edna.Domain modules and thus it makes more sense to have errors per
domain. There is no need to group them into one EdnaServerError type
because such grouping is already done by means of built-in exception
machinery.

Solution:
1. Create Edna.Upload.Error with all upload-specific errors.
2. Move EdnaServerError there and rename to UploadApiError.
3. Delete `XlsxParingError` because it's used only in a legacy
function and is not needed there.
4. Provide a default definition for `toServerError` to make it easier
to define its instances.
